[flang] Recognize target-specific tests correctly when using FLANG_TEST_TARGET_TRIPLE
FLANG_TEST_TARGET_TRIPLE is used to switch the target when running Flang regression tests,
but not used to determine which tests should be run.
This patch fixes it.

Note that the target for tests specified by FLANG_TEST_TARGET_TRIPLE can be overridden by
the value of the environment variable specified by LLVM_TARGET_TRIPLE_ENV.